integration: extract test_request_timeout()
The test is extracted to `statements/request_timeout.rs`.
