+++ scylla/tests/integration/statements/mod.rs
@@ -2,6 +2,7 @@ mod batch;
 mod consistency;
 mod execution_profiles;
 mod prepared;
+mod request_timeout;
 mod silent_prepare_query;
 mod skip_metadata_optimization;
 mod statement;
